remove redundant listeners
Set visibility to visible: We're explicitly making sure that the modal is visible when it's being shown, especially if there are CSS issues with visibility.

Click on the background to close: This logic ensures that clicking anywhere on the background of the modal (but not the image) will close the modal.

Click on the image to close: This feature makes it so that clicking the image itself will close the modal. This gives the user an easy and intuitive way to close it, just by clicking the image.

Locking and restoring scroll: When the modal is open, we lock the page scroll to avoid any unwanted page movement. Once the modal is closed, we restore the page scroll.

What's removed/modified:
Removed any extra logic that was redundant, like unnecessary event listeners that were not scoped properly.

Kept the event listeners scoped tightly to the modal (for improved performance).